+++ inst/tinytest/test_simulate.R
@@ -104,6 +104,11 @@ expect_silent(
       eta = cbind(etaLambda), 
       seed = 1
     ),
+    oichao = simulate(
+      oichao(),
+      eta = cbind(etaLambda),
+      seed = 1
+    ),
     zelterman = simulate(
       zelterman(), 
       eta = cbind(etaLambda), 
@@ -117,6 +122,10 @@ expect_true(
 )
 
 expect_true(
+  all(res$oichao %in% 2:3)
+)
+
+expect_true(
   all(res$zotpoisson == res$ztpoisson)
 )
 
@@ -127,4 +136,3 @@ expect_true(
 expect_true(
   all(res$zotnegbin == res$ztnegbin)
 )
-